asyncapi: white-box fitness + AsyncAPIProblem driver DTO (M5)
Lights up the white-box AsyncAPI flow.  After this PR a driver can
return AsyncAPIProblem from getProblemInfo() and the EvoMaster core
will run a Kafka request/reply search with branch/line coverage from
the EM Driver layered on top of the schema-derivable targets M4b
already produced.

Driver-side surface (client-java):
- AsyncAPIProblemDto in controller-api (carries asyncApiUrl, inline
  schema fallback, brokerBootstrapServers, correlationIdHeader).
- AsyncAPIProblem pojo in controller, mirroring the RestProblem
  contract (URL or inline schema, validating constructor).
- SutInfoDto.asyncAPIProblem field.
- EMController populates it when getProblemInfo() returns the new
  pojo, sitting next to the existing rest/graphql/rpc/web branches.

Core fitness refactor:
- New AbstractAsyncAPIFitness owns the broker round trip and all
  schema-derivable target tracking that M4b put in
  AsyncAPIBlackBoxFitness.  Two open hooks let subclasses extend the
  loop:
    * beforeAction(action, index) — white-box overrides to
      registerNewAction so EM Driver coverage attribution matches.
    * afterIndividualEvaluated(individual, fv, ...) — white-box
      overrides to call updateFitnessAfterEvaluation and merge in
      branch/line targets.
- AsyncAPIBlackBoxFitness is now a one-line subclass; behaviour
  unchanged from M4b.
- AsyncAPIFitness is the white-box leaf; uses the EnterpriseFitness
  helpers EvoMaster already provides.  Coverage poll happens once at
  end-of-individual rather than per-action because the broker hop is
  asynchronous; M6's hybrid wait strategy will refine this.
- AsyncAPISampler resolveSchemaUrl() now branches on blackBox: --
  blackBox reads --bbAsyncApiUrl as before; otherwise it asks the EM
  Driver for the URL via SutInfoDto.asyncAPIProblem.
- New AsyncAPIModule binds the white-box fitness, RemoteController,
  StandardMutator, and StructureMutator.
- Main.kt picks up info.asyncAPIProblem in the inference branch and
  routes to AsyncAPIModule when --blackBox is false.  The earlier
  "white-box not implemented" guard is gone.

Test plan:
- mvn -pl core test -Dtest='AsyncAPIAccessTest,EMConfigTest,
  RestActionBuilderV3Test' — 264/264 pass.
- mvn -DskipTests install (full reactor) — clean.
- The dataset's rest-kafka-ncs drivers compile against the new DTOs;
  the swap (RestProblem → AsyncAPIProblem) lands as a follow-up
  Dataset PR.
